ovpn3cli/GDBus++: Migrate 'openvpn3 log'
It required a bit more changes to migrate the log command.  The code has
been mostly cleaned up and uses a simpler design and better split
responsibilities.

The biggest difference is the use of the LogWriter class.  This enables
more streamlined logging and gives more features at the same time, like
the ability to have a more colourful logging.

All the code related to the logging feature is now under the
ovpn3cli::log namespace as well.

Signed-off-by: David Sommerseth <davids@openvpn.net>