ovpn3cli: Make main() provide argv[0] details in exceptions
When the main() function catches various exceptions, include the
basename of argv[0] to give better indication of which program is used.

The output will start with "PROGRAM/COMMAND:", where PROGRAM is the
basename of argv[0] and COMMAND is the command being called in the
program.  The '/' separator is already used in similar places in the
cmdargparser, so we keep the same style for consistency.

Signed-off-by: David Sommerseth <davids@openvpn.net>