ovpn3cli/log: LogAttach::lookup_config_name() does not always need iterations
The LogAttach::lookup_config_name() method will iterate several times to
try to catch the Log signals as early as possible if the session has not
yet been started.

This kind of iteration only needs to happen when this method is called
via the LogAttach::sessionmgr_event() method.  When called via
LogAttach::AttachByConfig(), if the config is present - it does not need
this multiple iteration step at all; everything is guaranteed to be
present.  And if the config isn't already present - it will wait for
it via LogAttach::sessionmgr_event().

Signed-off-by: David Sommerseth <davids@openvpn.net>