build: Remove the '-dev' extension to binaries
The migration is now basically complete. This steps finalizes the
migration.  All the compiled C++ binaries which was installed had a
'-dev' extension so it would be easy and quick to switch back in case of
errors and for debugging: which allowed the old stable release to be
installed in parallel.

The code now is considered stable and feature rich enough to become
the main installation.

Signed-off-by: David Sommerseth <davids@openvpn.net>