ovpn3cli: Fix failing shell-completion for 'openvpn3 log'
When providing shell-completion results for configuration names to be
used with the 'openvpn3 log' command, it would fail.

The reason is that both the arghelper_config_names_sessions() and
arghelper_config_names() functions creates their own DBus::Connection
object internally.  When this object exits its scope, the D-Bus
connection is closed and the socket used to communicate with the D-Bus
service comes unavailable.  This situation makes it impossible to
re-connect to the D-Bus, due to the main socket being unavailable.

When the next arghelper function for the 'openvpn3 log' command is
called, it first calls the arghelper_config_names() and then it calls
arghelper_config_names_sessions() where the last one tries to create a
new DBus::Connection object, which just fails.

To circumvent this situation, these two main arghelpers have been
duplicated with a sibling which takes a DBus::Connection object as
argument.  The original arghelper functions will create the
DBus::Connection object as before but call this new variant with
the DBus::Connection as an argument.

This allows the arghelper for the 'openvpn3 log' command to create a
shared DBus::Connection object to be used across these two arghelpers.

And while at this code, switch to use std::find_if() for the duplicate
check and simplify the check if the config name is empty.

Signed-off-by: David Sommerseth <davids@openvpn.net>