configmgr: Adjust ACLs for Fetch/FetchJSON and public_access rights
The Fetch and FetchJSON D-Bus methods was handled slightly differently.
Make them have the same access rights and restrict this to only owner
and users with granted access, unless locked down.  If locked down,
only the owner has access to these methods.  The OPENVPN_USERNAME will
always have access, as it needs that to start VPN profiles.

The other methods will be loosened a bit up, which will grant access
if public_access is allowed.

Signed-off-by: David Sommerseth <davids@openvpn.net>