configmgr/proxy: Add/Remove tag methods provided odd errors
When AddTag() or RemoveTag() calls resulted in an error, the
CfgMgrException error was less useful:

     [Configuration Manager]  some_tag

The expected error messages are these:

     Tag already exists, tag value: some_tag
     Non-existing tag, tag value: some_tag

This issue appeared due to the error messages provided via GDBus++ are
now differently formatted, and less "extraction magic" was needed.  The
old "magic" resulted in skipping the core of the error message in the
new format.

Signed-off-by: David Sommerseth <davids@openvpn.net>