ovpn3cli/config-manage: Extract all profile before displaying it
The 'openvpn3 config-manage --show' command would provide some odd
output if the calling user does not have access to a profile - and the
profile was attempted accessed via the --config-path argument.

By first extracting all the needed information, it's possible to detect
early if the calling use has access to the profile.  This can be used to
bailout early if no information could be extracted.

This also enables the printing of information to be done a bit simpler,
as all the early parsing of the information can format and prepare all
the wanted data.

Signed-off-by: David Sommerseth <davids@openvpn.net>