netcfg/proxy: Fix incorrect data type for Device::SetMtu()
This method used the incorrect data type (uint32_t) of what
the net.openvpn.v3.netcfg service expected (uint16_t) resulting in an
error when this was called.

The reason this was not discovered earlier was that GDBus++ proxy
code didn't wait for a response when calling the
org.freedesktop.DBus.Properties.Set method.  Since GDBus++ now
does this call synchronously, it will also catch these errors.

Signed-off-by: David Sommerseth <davids@openvpn.net>