configmgr: Fix mixed up Log Attach() API usage
The LogServiceProxy class provides two different ways to instantiate
the LogServiceProxy object.  One which just creates the object and
another one which does the Attach() call in addition.

Calling both LogServiceProxy::Create() and then the AttachInterface()
via the create object can result in unexpected behaviours, like
attempting to attach twice.

Also ensure the Service doesn't stop if this fails for now. Just log the
error.  If the Configuration Manager doesn't log everything in this
situation, that isn't the most critical situation at the moment.

Signed-off-by: David Sommerseth <davids@openvpn.net>