sessionmgr: Remove the backend VPN proxy if session closing fails
If the backend VPN client has stopped, the session manager will try to
remove the VPN backend in a loop despite the Disconnect/ForceShutdown
calls to the backend fails each time.  The close_session() method can be
called via the Authorize() method as well when checking situation before
authorizing access.

Also catch all the DBus::Exception() errors directly.  This ensures that
the backend proxy setup is guaranteed to be invalid if the
Disconnect/ForceShutdown fails - and the validate_vpn_backend() will
also help avoid attempting calling an unavailable backend.

Signed-off-by: David Sommerseth <davids@openvpn.net>