sessionmgr: Catch errors if DBus::Object::Manager::RemoveObject() fails
The SessionManager::Session::close_session() method will call the
DBus::Object::Manager::RemoveObject() to remove a session which is to be
closed.  But it may happen this has already been done in a parallel
thread, which causes this to fail.

For now, just add a Debug() logging of this scenario - to avoid a little
"splat" in the log.  This is a nicer behaviour and it's easier to debug
this later on if needed.  This scenario is not expected to be a common
scenario but may happen if the overall system is in an unexpected state.

Signed-off-by: David Sommerseth <davids@openvpn.net>