configmgr/proxy: Calling CheckObjectExists() should not happen in ctor
The OpenVPN3ConfigurationProxy constructor should not call
CheckObjectExists() and throw an exception.  The implementer should
instead call the CheckObjectExists() afterwards if it needs to verify
the object really exists.

This avoids throwing off an unrecoverable situation, where the object
might not yet have been created (but will slightly later).  For example,
the Configuration Manager might not have completed adding or loading
all configuration profiles yet; especially if it just started.

With this change, the implementer also gets better control how to handle
this situation.

Signed-off-by: David Sommerseth <davids@openvpn.net>