sessionmgr/proxy: Improve the SessionManager::Proxy::Manager setup
Add a variable setup delay, up to 750ms in 150ms increments, for
the Manager object.  This is to ensure the proxy object will have
access to a running Session Manager service.

This is to avoid errors if the Session Manager was not running and it
took a little bit time to get settled.

Signed-off-by: David Sommerseth <davids@openvpn.net>