configmgr: Ensure log service is available on config import
If the net.openvpn.v3.log service is unavailable, importing a new
configuration profile will fail.  This is a corner case, as it isn't
expected that the log service will just disappear.  However, the
Config Manager should still be able to complete the import.

If the log service still cannot be reached after attempting to start it
via the DBus::Proxy::Utils::DBusServiceQuery::CheckServiceAvail(), it
will need some more investigation on the system.  This service is
expected to be functional.

Signed-off-by: David Sommerseth <davids@openvpn.net>