client: Check vpnclient object exists in 'Ready' callback
There is no guarantee the vpnclient object exists.  The older
dynamic-challenge protocol would result in a full disconnect before the
client is re-started when providing the response to the challenge.

Without this check, the Ready method will fail retrieving the run status
of the VPN session, as the vpnclient member will be nullptr.

Signed-off-by: David Sommerseth <davids@openvpn.net>