ovpn3cli/session-start: Improve stability starting sessions from file
Using the 'openvpn3 session-start' where the '--config' argument points
at a file to use (not pre-imported configuration), it could often fail
with:

     $ openvpn3 session-start --config client.ovpn
     Using configuration profile from file: client.ovpn
     [DBus::Connection] Connection flush failed The connection is closed
     openvpn3/session-start: ** ERROR ** Could not start new VPN session: Failed to start new tunnel

On the following attempt, it always works - but it's a slight difference
in the output:

     $ openvpn3 session-start --config client.ovpn
     Using pre-loaded configuration profile 'client.ovpn'

The 'Using pre-loaded' indicates the configuration profile is already
pre-imported to the configuration manager.  That happened in the first
attempt, but when setting up the connection to the Session Manager it
would fail as the DBus::Connection object closed the socket to the D-Bus
daemon.  Since the first attempt imported the configuration profile with
the 'single-use' flag set, a third attempt would fail just as the first
attempted failed; there exists no pre-imported configuration on the
third-attempt.  The fourth attempt works again.

This is an issue which is the same which was fixed for the failing
shell-completion in the log command, in commit 8a13d1b8d88748aebd.  When
the import_config() function's DBus::Connection object exits its scope,
the socket to the D-Bus daemon.

The solution is to let the 'session-start' command use a shared
DBus::Connection object, for both the import_config() function and the
following connection to the Session Manager proxy object.

Signed-off-by: David Sommerseth <davids@openvpn.net>