configmgr: Extend configuration objects with a Validate() D-Bus method
This new method will currently do a very simple check to see if the
configuration profile has the bare minimum needed to start a connection.
It will also report an error if server locked profiles are attempted;
which is a old type of OpenVPN Access Server configurations which is no
longer supported by the clients.

Signed-off-by: David Sommerseth <davids@openvpn.net>