ovpn3cli/config: Add profile validity check in config-manage and configs-list
This extends the output to highlight if there a configuration profile is
not considered valid.

The config-manage command adds this as a separate warning line with the
reason.  The configs-list will prefix the profile name with '!!' and the
verbose output will also provide the reason.  The JSON output of
configs-list will have both a new 'valid' boolean key and
'invalid_reason' if the profile is not valid.

Signed-off-by: David Sommerseth <davids@openvpn.net>