netcfg/proxy: Check non-response calls for nullptr before freeing
The g_variant_unref() is unhappy if it receives a nullptr.  For D-Bus
calls which does not return any value, the GVariant * return value of
the Proxy::Client::Call() method may be nullptr.

This adds a simple check those places where this scenario may happen in
the netcfg proxy code.

Reported-by: Lev Stipakov <lev@openvpn.net>
Signed-off-by: David Sommerseth <davids@openvpn.net>