configmgr: Load configuration profiles before starting the D-Bus service
The current implementation would load all the already saved
configuration profiles into memory when the Configuration Manager is
registered unto the D-Bus.  If there is a lot of configuration profile
parsing needed, all profiles might not be loaded in time before D-Bus
proxy clients starts accessing this service.

Instead load everything when the --state-dir argument is parsed, which
happens before the glib2 main loop is started.  This ensures the service
is fully settled before it starts receiving requests over the D-Bus.

Signed-off-by: David Sommerseth <davids@openvpn.net>