build: Set PACKAGE_NAME to 'OpenVPN3/Linux'
All prior releases have used 'OpenVPN3/Linux' in PACKAGE_NAME but that
changed when switching to Meson, as that would use the
meson.project_name() value instead - which results in 'openvpn3-linux'.

Since this value will end up in the IV_GUI_VER, this can cause issues
on servers parsing that information.  Revert back to the old string to
avoid further issues.

Signed-off-by: David Sommerseth <davids@openvpn.net>