configmgr: Accumulate proxy feature flags instead of overwriting
Currently the code checks the product version, and based on that
it sets available features for the configuration manager proxy.

But it overrides them (with =) instead of accumulating them (with
operator |=).

This commit fixes the issue. It also corrects a tiny comment typo
and makes CfgMgrFeatures a "regular" (i.e. not a "class" enum,
since its members are being used for bit manipulation - so as
plain unsigned integers).

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>