netcfg: Check stub-resolv.conf before giving up on systemd-resolved
We are currently only checking for the special systemd 127.0.0.53
nameserver in /etc/resolv.conf. However, on some systems (e.g. on
Debian 12, when using systemd-networkd), this is set up in
/run/systemd/resolve/stub-resolv.conf. This patch checks that file
as well before deciding that systemd-resolved is not usable.

Without this patch, on affected systems we try to work directly
with the /etc/resolv.conf file, which leads to the added
information not being removed on VPN connection shutdown.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>