netcfg/resolved: Add new D-Bus IP Address parser class
The new NetCfg::DNS::resolved::IPAddress class implements
a simpler interface to send and receive IP addresses when
communicating with the systemd-resolved service.

The systemd-resolved D-Bus API requires IP addresses to
be sent and received as an array of bytes together with
an address family indicator.

This implementation requires a GDBus++ version supporting
the D-Bus BYTE ('y') data type.

Signed-off-by: David Sommerseth <davids@openvpn.net>