configmgr/overrides: Remove struct OverrideValue
OverrideValue is a struct that partly acts as a discriminated
union with a bool and string option (which std::variant already
offers), and partly "connects" with ValidOverride by actually
storing (by-value) a "parent" ValidOverride member.

This patch removes OverrideValue and moves the value as an
std::variant into ValidOverride, then uses ValidOverride
everywhere.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>