client: Add support for --dhcp-option ADAPTER_DOMAIN_SUFFIX
The ADAPTER_DOMAIN_SUFFIX is a slightly different approach of
setting a search domain in OpenVPN.  Strictly speaking, it is
a Windows specific feature which is used by the Windows network
stack - which implies setting a single DNS search domain.  The
behaviour of also setting the DNS search domain is expected on
other platforms as well by users.

Signed-off-by: David Sommerseth <davids@openvpn.net>