client: Stop running VPN clients in client destructor
If the openvpn3-service-client process is receiving a SIGINT/SIGTERM
signal with an on-going VPN session running, it will hang in the
~BackendClientObject() destructor.

The destructor will wait for the client thread to complete first, but
that thread as not been told to stop.  Instead ensure that the client
thread is being stopped before doing the join() call on the thread.

Signed-off-by: David Sommerseth <davids@openvpn.net>