build: reduce hardcoded 'asio_path'
Currently, 'asio_path` variable value is concatenated with
'/asio/include' to specify the path to custom `asio`
installation.  The problem is that this is too strict as some
distros (namely NixOS) may have the 'include' directory with
a differently named parent.  Thus this change minimizes the
hardcoded part of the path to make it more flexible.

Signed-off-by: Petr Portnov <mrjarviscraft@gmail.com>