client: Properly plug-in DBus::MainLoop handling in BackendClientObject
The D-Bus Disconnect method did not properly stop the process.  One
issue was that it expects a valid vpncli_ptr pointer; which may not
always be true - resulting in a segfault.

Another issue was that it did not properly stopping the DBus::MainLoop
of the main process.  The process used mainloop provided via
DBus::Service - but that was never passed on to the BackendClientObject.
It turned out the DBus::MainLoop setup was also not done at all via the
ClientService object.

This has been refactored and a DBus::Mainloop object is created before
the ClientService object and the mainloop object is passed to it, which
will pass it further to the BackendClientObject when that object is being
set up.

This allows the Disconnect (via cb_disconnect()) to stop the mainloop
correctly.  And also this method needed some cleanup as well.  It is
irrelevant if the client is registered or not.  Just shut down
regardless.

Signed-off-by: David Sommerseth <davids@openvpn.net>