client: Make BackendSignals::LogFATAL() thread safe
It could be that different threads would call the LogFATAL() method,
which would remove the existing delayed_shutdown thread object.  Leave
the existing object running and skip creating a new one.

Also switch out the sleep() with a C++ approach instead.  The sleep(3)
man page claims it is not thread-safe, which is hard to believe (since
it is calling nanosleep() internally).  But using a more native C++
approach seems better in this context.

Signed-off-by: David Sommerseth <davids@openvpn.net>