configmgr: Add debug option --use-session-bus
This is a debug/test feature to start the Configuration Manager on the
session D-Bus instead of the system D-Bus.  For this to work, the
openvpn3-service-log must have been started in advance on the session
bus as well.

The config-override-selftest will now also take the --use-session-bus
argument to connect to the Configuration Manager on the session bus.

Signed-off-by: David Sommerseth <davids@openvpn.net>