log: Rework the tear-down of ProxyLogEvents objects
The current implementation gets use-after-free/double-free-like
situations in the AttachRemoveCallback() lambda method when
ProxyLogEvents objects are removed - which causes the log service
to segfault and crash.

The reason is due to the ~ProxyLogEvents() destructor will implicitly
trigger the AttachRemoveCallback() which tries to access the
ProxyLogEvents object no longer available.  In addition, there is a risk
that the AttachedService() access to the proxies vector will have issues
in the various for-loops if multiple callers accesses this object.

The initial purpose of the AttachRemoveCallback() lambda was to ensure
no stray ProxyLogEvents() objects was left dangling behind and log the
removal.  The smart-pointers should take care of the dangling objects
situation; the reference counter clearly gets to zero due to the
segfault.  And the logging can be done in the ~ProxyLogEvents()
destructor instead.

Signed-off-by: David Sommerseth <davids@openvpn.net>