sessionmgr:  Implement extraction of connection details from client
Client services (net.openvpn.v3.backends/openvpn3-service-client) which
exposes the 'connection' property will now take some of that information
and expose that as the 'connected_to' property in the Session Manager
objects.  This new property has the data format '(ssu)':

   s - protocol (udp/tcp, udp-dco, tcp-dco)
   s - server IP address
   u - server port used for the connection

If this information is unavailable, the string elements in the response
will be empty and the port number will be 0.

Signed-off-by: David Sommerseth <davids@openvpn.net>