ovpn3cli/sessions-list: Improve session information with connection details
The prior implementation would present the "Session name" line, which
often would contain the server name of the VPN server the client is
connected to.  If a configuration profile contained more --remote
options with different names, this would not be reflected if the client
connects to one of the other servers.

The backend VPN client and session manager services can now provide more
accurate connection information and when available, it will show this
information as a "Connected to:" line instead of the "Session name:"
line.

The explicit DCO flag has been removed, as that is now reflected in the
protocol line of the new connection details line.

Signed-off-by: David Sommerseth <davids@openvpn.net>