netcfg/proxy: Extend NetCfgProxy::Device with openvpn::DnsOptions parsing
The configuration of DNS resolver settings has so far happened via the
AddDNS(), AddDNSSearch() and SetDNSscope() methods.  In this change, the
NetCfgProxy::Device class is extended with AddDnsOptions() which takes
an openvpn::DnsOptions object containing the --dns options parsed by the
OpenVPN 3 Core library.

In addition, this method will take a LogSender object which will be used
for logging when the processing is having some issues.

Signed-off-by: David Sommerseth <davids@openvpn.net>