netcfg/systemd-resolved: Implement support for setting the DNSSEC mode
The netcfg service is being extended to support the new --dns option
which can control the DNSSEC modes.  Three modes are available:

   yes      - DNSSEC validation is mandatory
   no       - DNSSEC validation is disabled
   optional - Opportunistic DNSSEC validate; will not hard-fail on
              incorrect validations

The default is "unset", which means it will use the default setting of
the systemd-resolved.

This change will only implement the capability for the
openvpn3-servivce-netcfg service to configure DNSSEC in systemd-resolved
"under the hood".  This feature is not yet exposed.

Signed-off-by: David Sommerseth <davids@openvpn.net>