netcfg/systemd-resolved: Implement support for setting the DNS transport mode
The netcfg service is being extended to support the new --dns option
allows to control the DNS protocol used to query DNS servers.  The
supported modes depends on what the resolver backend supports, but the
general implementation is prepared for the following modes:

     - plain  -  Unencrypted (traditional port 53)
     - dot    -  DNS over TLS (encrypted)
     - doh    -  DNS over HTTPS
     - unset  -  Not set, use the default of the backend resolver

The default is "unset", which means it will use the default setting of
the systemd-resolved.

For systemd-resolved, only "plain" and "dot" are supported.  Whether
"dot" will be fully enforced also depends on the systemd-resolved
version.  Version 239 will only do an "opportunistic" approach while
v245 can enforce "dot".

At this stage, it's not possible to set the "sni" of the DNS server -
which will be important to reduce the possibility of man-in-the-middle
attacks on the DNS queries.  Support for setting this will arrive later.

This change will only implement the capability for the
openvpn3-servivce-netcfg service to configure DNSSEC in systemd-resolved
"under the hood".  This feature is not yet exposed to any external
users.

Signed-off-by: David Sommerseth <davids@openvpn.net>