netcfg/systemd-resolved: Refactor and simplify the code
The SystemdResolved::Commit() method which will make the requested
changes effective began to be hard to read.  This refactors and
restructures the code a bit by adding helper functions and
reordering some of the if() logic.  This also reduces the lines
of code and the local scopes in this method.

Signed-off-by: David Sommerseth <davids@openvpn.net>