sessionmgr: Allow a Session object to re-associate with a backend process
Expose Session::ResetBackend() for the purpose of allowing a
session object to link to another backend process. This is useful
in case the current backend process dies and we want to restart
it. ResetBackend() will create a new D-Bus proxy to the new
backend process and set up the signals infrastructure for it.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>