sessionmgr: Restart prematurely stopped backend processes
This patch immediately restarts and backend process that has ended
before it's corresponding Session is done, if the
"automatic-restart" configuration override is set to "on-failure".

This is the simplest possible way to do that, so it is more of a
PoC than anything. Meaning, it always immediately tries to restart
and does so forever, which is not particularly sophisticated.

Future work may allow for only a limited number of restarts at
increasingly larger intervals.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>