sessionmgr/proxy: Replace sleep with waiting for SESS_CREATED signal
The current implementation of the SessionManager::NewTunnel() method was
very simplistic and could result in incorrect behaviour, due to the use of
a simple sleep(1).  Even though 1 second might in most cases be enough
to start a new VPN client process, in some systems it might take
slightly longer.

Since the Session Manager service will emit a SessionManagerEvent signal
with the SESS_CREATED type, rework this code to wait for this signal
before continuing.  If this signal is not received within 15 seconds, it
will fail with an exception to the caller.

Signed-off-by: David Sommerseth <davids@openvpn.net>