sessionmgr: Fix misbehaviour if GetUID() fails in Session::helper_stop_log_forwards()
In some cases a process which has requested log forwarding on a VPN
session may not have stopped the forwarding before closing.  This
results in the DBus::Credentials::Query::GetUID() to fail, as it cannot
lookup the UID to the owner of a bus name since that bus name is no
longer available.

Capture this situation and let the clean-up continue unaffected.  The
information attempted to be retrieved is only used for more logging
details.

Signed-off-by: David Sommerseth <davids@openvpn.net>