ovpn3cli: Improve mainloop start synchronisation
In commit e51c46fc9ed57cd, a glib2 main loop is now started for all
commands, regardless if it needs it or not.  But to avoid some odd
behaviours a quick hack was added using a 25ms sleep to ensure the
main loop had truly started.

This had the negative side-effect of often losing some log and status
change events in the 'openvpn3 log' command when the log command was
started about the same time as the session started.

This change replaces the 25ms sleep with a conditional variable
synchronisation between the thread running the glib2 main loop and the
thread running the main() function.

To accomplish this, a function is now once run via a g_idle_add()
callback.  This approach ensures that the callback function is not run
before the glib2 main loop is really running.  And that callback can
then release the conditional variable locking.

A simple tests measuring the time needed to run 'openvpn3 version'
indicates the spent time is reduced from a span between 32ms to 56ms
down to 7ms to 16ms.

Signed-off-by: David Sommerseth <davids@openvpn.net>