common: Check if org.freedesktop.hostname1 is available in PlatformInfo
On systems without systemd available, the new setup logic in the
DBus::Proxy::Client will result in a slow failure when trying to
access a non-existing service.  This causes openvpn3-service-client
to timeout and the session is torn down by the session manager before
it gets a chance to start.

This implements a quick check if this service is available.  If not,
just bail out quickly without trying to setup the DBus::Proxy::Client.

Signed-off-by: David Sommerseth <davids@openvpn.net>