ovpn3cli/admin: Improve sessionmgr-service verose session list
The --list-sesisons-verbose mode of the sessionmgr-service command
reported the session name and not details about the real connected
server.

The session name field is now removed and the session status has
been moved up to this place.  The prior session status line is
now used to provide details about the currently connected server.

Signed-off-by: David Sommerseth <davids@openvpn.net>