client: Refactor D-Bus initialization during process start
The client process would often send the "RegistrationRequest" signal at
a very early stage, where the main BackendClientObject might not be fully
ready to accept D-Bus calls in some cases.  This would typically make
the Session Manager shutdown the session due to the client process not
being ready - while indicating it was available at the same time.

This refactoring moves the BackendClientObject setup very early, before
it completes the registration unto the D-Bus.  And it will only send the
RegistrationRequest signal once it has been given a bus name.

Signed-off-by: David Sommerseth <davids@openvpn.net>