netcfg/resolved: Add internal error message storage to proxy code
This implements a simple storage for keeping errors messages per link
interface.  This will be used when extending the systemd-resolved D-Bus
proxy with support for asynchronous background calls.

The current implementation could have had a middle layer between the
main storage container and the struct holding a single error message.
This middle layer would then keep and handle messages for a single link
interface.  But it was chosen to use a more simpler approach as the use
case is highly restricted to this proxy code.

Signed-off-by: David Sommerseth <davids@openvpn.net>