netcfg/resolved: Implement base features for background async calls
The functionality added here should not change the current behaviour
yet, but provides the framework to do background D-Bus calls and be able
to retrive errors which might have happened.

These background calls will not be able to provide any results back to
the caller.  This is to be used where the caller does not need to
retrieve a result and the call is not time-critical.

Signed-off-by: David Sommerseth <davids@openvpn.net>