netcfg: Refactor D-Bus initialization during process start
The openpvn3-service-netcfg service does not always respond quickly
enough to D-Bus calls after it has been registered unto the D-Bus.  This
is due to the "service handler" first being initialized after the
well-known bus name has been acquired.

This change moves all of the initialization to happen before it is being
registered on the bus.  This way it will be able to respond instantly
once it is being announced as available on the D-Bus.

Signed-off-by: David Sommerseth <davids@openvpn.net>