sessionmgr: Ignore Detach() exceptions in SessionManager::~Service()
Ignore exceptions from the Detach() call in the destructor for
SessionManager::Service().  When this happens, it's because the log
service does not recongise this service any longer.

Since the Session Manager is already shutting down and it's the very end
of the shutdown process, ignoring this error is perfectly fine.

Signed-off-by: David Sommerseth <davids@openvpn.net>