ovpn3cli/init-config: Refactor file/directory handling to use std::filesystem
The std::filesystem provides a simpler API compared to the lower level
C APIs for creating directories and setting permissions, as well as some
file checks.  This also removes the dependency on the OpenVPN 3 Core
library's common/stat.hpp code.  The only thing std::filesystem lacks
is the chown() function.

Signed-off-by: David Sommerseth <davids@openvpn.net>