ovpn3cli/init-config: Don't follow symlinks setting up state/configs dirs
The openvpn3-admin init-config command will attempt to ensure the
openvpn3 state directory and the configs sub-directory is properly
configured with both permissions and ownership.

What the code didn't account for was that it would follow symlinks as
well, which could be abused in various attack vectors.

To avoid this, the state/configs directories setup will be skipped if a
symlink is detected on either of these directories.

CVE: CVE-2025-3908
Reported-by: Wolfgang Frisch <wfrisch@suse.de>
Signed-off-by: David Sommerseth <davids@openvpn.net>