sessionmgr: Catch incorrect log level requests in Session object
When the Session object is requested to change the log level
(log_verbosity) to an invalid log level, a LogException is thrown.  This
exception is not caught, thus it will crash the Session Manager.

By rethrowing this exception as a DBus exception, the GDBus++ layer will
propagate this error correctly back to the end-user.

Reported-By: Wolfgang Frisch <wfrisch@suse.de>
Signed-off-by: David Sommerseth <davids@openvpn.net>
(cherry picked from commit a5ebf1ca8a72536f70df1de49a9ddcc28dc4ac52)