events/log: Refactor Events::Log()
Simplify the Events::Log struct by removing some of the constructors
out of the main class and into separate Events::ParseLog() functions
returning Events::Log() objects instead.

This goal is to make the Events::Log API a bit simpler and move the
lesser used string to LogGroup/LogCategory and GVariant to Events::Log
parsing out of the core of the Events::Log.

Signed-off-by: David Sommerseth <davids@openvpn.net>