build: Improve OpenVPN 3 Core library version extraction
The script/version tool in the Core library is fairly simple and will
always give a git commit reference.  The script/get-version tool in the
openvpn3-linux project is capable of extracting version tags if
available.

By slightly modifying the get-version script, it's possible to extract
more user friendly core version.  This is put into a separate mode, by
giving the --core-version argument instead.  The regex needed to detect
the version tags is slightly different and the .git directory to use is
in a different location.  Otherwise, most of the code can be reused.

Signed-off-by: David Sommerseth <davids@openvpn.net>