common: Filter input value to RequiresQueue::UpdateEntry()
The RequiresQueue::UpdateEntry() method receives data from an end-user.
It is not expected that this string contains any control or newline
characters, so filter it out.

The related test program is also extended to pass some control
characters when providing a value back to the RequiresQueue.  It is
expected that these characters are removed when checking the value later
on.

Signed-off-by: David Sommerseth <davids@openvpn.net>
(cherry picked from commit 00dc88631e3d74b4f7c9a2f0d4ef1dccebba0c44)