addons/aws: Implement support for additional route table
When config contains

  "route-table-name": "OpenVPN3-Linux"

the AWS addon will try to find the route table with "Name" tag
and the value "OpenVPN3-Linux". If the route table doesn't
exist, it will be created and "Name" tag will be assigned.
The routes will be then added to this route table instead of
subnet's main route table.

Signed-off-by: Lev Stipakov <lev@openvpn.net>