log/journald: Fix wrong timezone/dst handling in journald filter
The systemd-journald filter setup failed when requesting a timestamp
including hours.  It would not properly account for timezone or DST
differences.

The needed information is being retrieved using the localtime()
C library function.  So a little refactoring allows us to preserve
this information when parsing the user string into the std::tm object.

Signed-off-by: David Sommerseth <davids@openvpn.net>