log/journald: Refactor log event sending with better error handling
The JournaldWriter::Write() method is preparing several struct iov
object in a C array.  A lot of the needed steps are duplicated and
they were lacking proper error checking if the needed memory
allocation was successful.

This refactoring moves all the memory handling and struct iov
population logic into a local static function which will throw
an exception if memory allocation fails.

Signed-off-by: David Sommerseth <davids@openvpn.net>