build: Avoid including build-config.h in header files
Due to the intricacies with the header file inclusion order expected by
the OpenVPN 3 Core Library, the inclusion of build-config.h in some of
the OpenVPN 3 Linux header files can cause unexpected build issues.

These issues are typically avoided by including the build-config.h only
in the source files which contains the implementation of functions.

Signed-off-by: David Sommerseth <davids@openvpn.net>