netcfg/dns/systemd-resolved: Rework the resolved::Link::BackgroundCall() implementation
The current implementation had some issues which would often be caused
by the DBus::Proxy::TargetPreset object being used for storing the
D-Bus object path details to the systemd-resolved object was
unexpectedly expired.  This again resulted in the systemd-resolved DNS
settings to be thrown away.

Instead, implement a crude raw-pointer model for passing all the
needed details to the worker thread which will operate on this
information.

This approach is far from ideal.  But it seems to currently be more
reliable than the alternatives.  This code path should be revisited
later on.

Signed-off-by: David Sommerseth <davids@openvpn.net>